nested loops projects

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,12 +7,12 @@
 int main(void)
 {
 	int i;
-	int fib1 = 0, fib2 = 1, sum;
+	unsigned long fib1 = 0, fib2 = 1, sum;
 
 	for (i = 0; i < 50; i++)
 	{
 		sum = fib1 + fib2;
-		printf("%d", sum);
+		printf("%lu", sum);
 
 		fib1 = fib2;
 		fib2 = sum;
